% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_thermalmodel.R
\name{run_thermalmodel}
\alias{run_thermalmodel}
\title{Runs 1D integral energy model}
\usage{
run_thermalmodel(
  u,
  startTime,
  endTime,
  ice = FALSE,
  Hi = 0,
  iceT = 6,
  supercooled = 0,
  scheme = "implicit",
  kd_light = NULL,
  densThresh = 0.001,
  albedo = 0.1,
  eps = 0.97,
  emissivity = 0.97,
  sigma = 5.67 * 10^(-8),
  p2 = 1,
  B = 0.61,
  g = 9.81,
  Cd = 0.0013,
  sw_factor = 1,
  meltP = 1,
  dt_iceon_avg = 0.8,
  Hgeo = 0.1,
  KEice = 0,
  Ice_min = 0.1,
  area,
  depth,
  volume,
  zmax,
  nx,
  dt,
  dx,
  daily_meteo,
  Cdplant = 0,
  ahat = 0.4,
  km = 0,
  rho_plant = 30,
  canpy = NULL,
  biomass = NULL,
  water.quality = FALSE,
  wq = NULL,
  agents = NULL,
  Hs = 0,
  rho_snow = 250,
  Hsi = 0,
  rho_ice = 910,
  rho_fw = 1000,
  rho_new_snow = 250,
  rho_max_snow = 450,
  K_ice = 2.1,
  Cw = 4180000,
  L_ice = 333500,
  kd_snow = 0.9,
  kd_ice = 0.7,
  pgdl_mode = false
)
}
\arguments{
\item{u}{vector; initial temperature profile}

\item{startTime}{integer; starting time inseconds of the meteorology driver data}

\item{endTime}{integer; starting time inseconds of the meteorology driver data}

\item{ice}{logical; determines if ice is existing or not at model start. Defaults to FALSE}

\item{Hi}{double; ice thickness (m) at model start. Defaults to 0}

\item{iceT}{double; moving average of air temperature for ice formation. Defaults to 6 deg C}

\item{supercooled}{integer; amount of layers below 0 deg C at model start. Defaults to 0}

\item{scheme}{character; numerical diffusion scheme, either 'explicit' or 'implicit'}

\item{kd_light}{double; if NULL then the dataframe from daily_meteo is used, otherwise a DOUBLE value for light extinction (m-1)}

\item{densThresh}{double; density difference threshold between neighboring layers for convective overturn. Defaults to 1E-3}

\item{albedo}{double; water albedo. Defaults to 0.1}

\item{eps}{double; epsilon. Defaults to 0.97}

\item{emissivity}{double; water emissivity. Defaults to 0.97}

\item{sigma}{double; constant sigma value. Defaults to 5.67 * 10^(-8)}

\item{p2}{double; calibration multiplier. Defaults to 1.0}

\item{B}{double; constant. Defaults to 0.61}

\item{g}{double; gravity. Defaults to 9.81 m/s2}

\item{Cd}{double; momentum drag coefficient. Defaults to 0.0013}

\item{sw_factor}{double; calibration multiplier for short-wave radiation. Defaults to 1.0}

\item{meltP}{double; calibration multiplier for ice melting. Defaults to 1.0}

\item{dt_iceon_avg}{double; determines weighting of ice onset. Defaults to 0.8}

\item{Hgeo}{double; geothermal heat flux. Defaults to 0.1 W/m2}

\item{KEice}{double; TKE multiplier due to ice cover. Defaults to 0}

\item{Ice_min}{double; minimum ice thickness in m. Defaults to 0.1}

\item{area}{vector; area over depth in m2}

\item{depth}{vector; depths in m references from surface}

\item{volume}{vector; volumes over depth in m3}

\item{zmax}{double; maximum lake depth in m}

\item{nx}{double; maximum number of grid cells}

\item{dt}{double; model time step in seconds}

\item{dx}{double; model spatial discretisation in m}

\item{daily_meteo}{dataframe; meteorological driver data}

\item{Cdplant}{double; drag coefficient macrophytes. Defaults to 0.0}

\item{ahat}{double; surface to volume ratio macrophytes. Defaults to 0.4}

\item{km}{double; ight extinction macrophytes in m-1. Defaults to 0.0}

\item{rho_plant}{double; density macrophytes. Defaults to 30 kg/m3}

\item{canpy}{dataframe; canophy height time series, NULL refers to no macrophyte simulations. Defaults to NULL}

\item{biomass}{dataframe; biomass time series, NULL refers to no macrophyte simulations. Defaults to NULL}

\item{water.quality}{logical; if TRUE runs water quality (need a defined function 'water_quality_boundary_conditions'). Defaults to FALSE}

\item{wq}{vector; initial vector of water quality variable profile, needs water.quality = TRUE. Defaults to NULL}

\item{agents}{vector; initial vector of individual phytoplankton cells. Defaults to NULL}

\item{Hs}{double; snow thickness (m) at model start. Defaults to 0}

\item{rho_snow}{double; snow density (kg/m3). Defaults to 250}

\item{Hsi}{double; snow-ice thickness (m) at model starts. Default to 0}

\item{rho_ice}{double; ice density (kg/m3). Defaults to 910}

\item{rho_fw}{double; freshwater density (kg/m3). Defaults to 1000}

\item{rho_new_snow}{double; new snow density (kg/m3). Defaults to 250}

\item{rho_max_snow}{double; maximum snow density (kg/m3). Defaults to 450}

\item{K_ice}{double; thermal conductivity of ice (W/K/m). Defaults to 2.1}

\item{Cw}{double; volumetric heat capacity of water (J/K/m3). Defaults to 4.18E6}

\item{L_ice}{double; latent heat of freezing (J/kg). Defaults to 333500}

\item{kd_snow}{double; snow albedo. Default to 0.9}

\item{kd_ice}{double; ice albedoa. Defaults to 0.7}

\item{pgdl_model}{logical; get additional diagnostics. Defaults to false}
}
\description{
Returns a data frame containing model output in the form of vertical temperature
profiles, diffusivity values, mixing enery, buoyancy frequencies, ice thickness,
logical value for ice, ice moving average temperature, depth of supercooled
layers, mixing depth, thermocline depth, end time of model run, averaged
temperatures for epilimnion and hypolimnion volumes, incoming short-wave
radiation, incoming long-wave radiation, backscattered long-wave radiation,
latent and sensible heat fluxes.
}
\details{
The mixing algorithms are based on the MINLAKE (Ford and Stefan 1980, Riley
and Stefan 1988, Herb and Stefan 2004) and the MyLake (Saloranta and Andersen 2007)
models. Implementations to estimate the incoming and outgoing long-wave heat fluxes
were taken from Livingstone and Imboden (1989) and Goudsmit et al. (2002). The latent
and sensible heat fluxes were calculated taking into account atmospheric stability using
the algorithms by Verburg and Antenucci (2010). The ice algorithms from MyLake (Saloranta
and Andersen 2007) were applied to simulate ice formation and melting.

References:
Ford, D.E., and H.G. Stefan. 1980. Thermal predictions using an integral energy model. J. Hydraul. Div. ASCE 106(1). 39-55

Goudsmit, G.H., H. Burchard, F. Peeters, and A. Wüst. 2002. Application of k-e turbulence models to enclosed basins: The role of internal seiches. Journal of Geophysical Research 107. C12. 3230. doi:10.1029/2001JC000954

Herb, W.R., and H.G. Stefan. 2004. Temperature Stratification and Mixing Dynamics in a Shallow Lake With Submersed Macrophytes. Lake and Reservoir Management 20. 4. 296-308. doi:10.1080/07438140409354159

Livingstone, D., and D. Imboden. 1989. Annual heat balance and equilibrium temperature of Lake Aegeri, Switzerland. Aquat. Sci. 51

Riley, M., and H.G. Stefan. 1988. MINLAKE: A dynamic lake water quality simulation model. Ecol. Model. 43. 155-182

Saloranta, T.M., and T. Andersen. 2007. MyLake – A multi-year lake simulation model code suitable for uncertainty and sensitivity analysis simulation. Ecol. Model. 207. 1. 45-60. doi:https://doi.org/10.1016/j.ecolmodel.2007.03.018

Verburg, P., and J.P. Antenucci. 2010. Persistent unstable atmospheric boundary layer enhances sensible and latent heat loss in a tropical great lake: Lake Tanganyika. Journal of Geophysical Research Atmospheres 115. D11. doi:https://doi.org/10.1029/2009JD012839
}
\examples{
\dontrun{
res <-  run_thermalmodel(u = u_ini,
startTime = startTime,
endTime =  endTime,
ice = FALSE,
Hi = 0,
iceT = 6,
supercooled = 0,
kd_light = 0.5,
sw_factor = 1.0,
zmax = zmax,
nx = nx,
dt = dt,
dx = dx,
area = hyps_all[[1]], # area
depth = hyps_all[[2]], # depth
volume = hyps_all[[3]], # volume
daily_meteo = meteo,
Cd = 0.0013,
scheme = 'implicit')
}

}
\author{
Robert Ladwig
}
